backend/requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
Babel==2.9.1
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
dnspython==2.1.0
drf-yasg==1.20.0
email-validator==1.1.3
Faker==12.0.0
Flask==2.0.2
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
graphene==3.0
graphene-django==3.0.0b7
graphql-core==3.1.7
graphql-relay==3.1.5
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
mixer==7.2.0
packaging==21.3
passlib==1.7.4
promise==2.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.29
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
